[refactor] Dao 리팩토링
