+++ src/test/java/chess/player/PlayerTest.java
@@ -2,17 +2,11 @@ package chess.player;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import spring.chess.board.ChessBoard;
-import spring.chess.board.ChessBoardCreater;
-import spring.chess.location.Location;
-import spring.chess.piece.type.Piece;
 import spring.chess.player.ChessSet;
 import spring.chess.player.Player;
-import spring.chess.score.Score;
 import spring.chess.team.Team;
 
 import java.util.Collections;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -49,14 +43,4 @@ class PlayerTest {
 
         assertThat(player.isNotSame(Team.WHITE)).isFalse();
     }
-
-    @Test
-    void calculateScoreExceptPawnReduce() {
-        ChessBoard chessBoard = ChessBoardCreater.create();
-        Map<Location, Piece> locationPieceMap = chessBoard.giveMyPieces(Team.WHITE);
-        Player player = new Player(new ChessSet(locationPieceMap), Team.WHITE);
-
-        assertThat(player.calculateScoreExceptPawnReduce()).isEqualTo(new Score(38));
-    }
-
 }
